docs(readme): Relocate 'Buy Me a Coffee' link
Moved the "Buy Me a Coffee" badge and link from the top of the
README to a new, dedicated '## Support' section at the bottom.

This change aims to improve the visual presentation and organization
of the README, giving better visibility to support options while
decluttering the introductory section. The link is also centered.

Signed-off-by: Enric Balletbo i Serra <eballetbo@gmail.com>